+++ CHANGELOG.md
@@ -2,6 +2,40 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.5.13"></a>
+
+## [0.5.13](https://github.com/UCloud-FE/react-components/compare/v0.5.12...v0.5.13) (2019-09-02)
+
+### Bug Fixes
+
+*   **Message:** fix message onClose not trigger bug ([8bba21b](https://github.com/UCloud-FE/react-components/commit/8bba21b)), closes [#168](https://github.com/UCloud-FE/react-components/issues/168)
+
+### Features
+
+*   **Checkbox:** add support of multiple content ([eac8321](https://github.com/UCloud-FE/react-components/commit/eac8321)), closes [#170](https://github.com/UCloud-FE/react-components/issues/170)
+*   **Pagination:** add showTotal ([94f1112](https://github.com/UCloud-FE/react-components/commit/94f1112))
+
+<a name="0.5.12"></a>
+
+## [0.5.12](https://github.com/UCloud-FE/react-components/compare/v0.5.11...v0.5.12) (2019-08-22)
+
+### Bug Fixes
+
+*   **Dependences:** fix rc-\* dependences versions ([c765784](https://github.com/UCloud-FE/react-components/commit/c765784))
+
+<a name="0.5.11"></a>
+
+## [0.5.11](https://github.com/UCloud-FE/react-components/compare/v0.5.10...v0.5.11) (2019-08-14)
+
+### Bug Fixes
+
+*   **Popover:** fix popover scroll bar blur when scroll ([ba2b24f](https://github.com/UCloud-FE/react-components/commit/ba2b24f)), closes [#163](https://github.com/UCloud-FE/react-components/issues/163)
+
+### Features
+
+*   **Table:** add column.renderTitle & some docs ([b2767ff](https://github.com/UCloud-FE/react-components/commit/b2767ff)), closes [#164](https://github.com/UCloud-FE/react-components/issues/164)
+*   **Table:** add onConditionChange & doNotHandleConditionChange ([14ff4c5](https://github.com/UCloud-FE/react-components/commit/14ff4c5))
+
 <a name="0.5.10"></a>
 
 ## [0.5.10](https://github.com/UCloud-FE/react-components/compare/v0.5.9...v0.5.10) (2019-07-18)